requirements.txt
asgiref==3.5.2
Django==4.0.5
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
drf-nested-routers==0.93.4
django-environ==0.9.0
django-filter==22.1
django-rest-swagger==2.2.0
drf-spectacular==0.22.1
drf-yasg==1.20.0
django-injector==0.3.1
pytz==2022.1
sqlparse==0.4.2
tzdata==2022.1
psycopg2-binary==2.9.3
django-prometheus==2.2.0